Additional Test Fixes
