+++ build.gradle
@@ -293,8 +293,6 @@ dependencies {
   implementation group: 'org.apache.commons', name: 'commons-collections4', version: '4.4'
   implementation group: 'commons-validator', name: 'commons-validator', version: '1.7'
   implementation group: 'com.github.ben-manes.caffeine', name: 'caffeine', version: '3.1.8'
-  implementation group: 'io.vavr', name: 'vavr', version: '0.9.0'
-
 
   annotationProcessor group: 'org.projectlombok', name: 'lombok', version: versions.lombok
