+++ src/main/java/uk/gov/hmcts/divorce/document/content/JudicialSeparationCoRefusalTemplateContent.java
@@ -5,17 +5,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.document.CaseDataDocumentService;
-import uk.gov.hmcts.divorce.document.model.DocumentType;
 
 import java.time.Clock;
 import java.time.LocalDate;
 import java.util.Map;
 
-import static java.time.LocalDateTime.now;
-import static uk.gov.hmcts.divorce.caseworker.service.task.util.FileNameUtil.formatDocumentName;
 import static uk.gov.hmcts.divorce.divorcecase.model.LanguagePreference.WELSH;
-import static uk.gov.hmcts.divorce.document.DocumentConstants.REJECTED_REFUSAL_ORDER_COVER_LETTER_DOCUMENT_NAME;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.APPLICANT_1_FULL_NAME;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.APPLICANT_1_SOLICITOR_NAME;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.APPLICANT_2_FULL_NAME;
@@ -46,9 +41,6 @@ public class JudicialSeparationCoRefusalTemplateContent {
     private static final String IS_JOINT = "isJoint";
 
     @Autowired
-    private CaseDataDocumentService caseDataDocumentService;
-
-    @Autowired
     private Clock clock;
 
     @Autowired
@@ -56,53 +48,6 @@ public class JudicialSeparationCoRefusalTemplateContent {
 
     @Autowired
     private ConditionalOrderCommonContent conditionalOrderCommonContent;
-
-    public void generateAndUpdateCaseData(final CaseData caseData,
-                                          final Long caseId,
-                                          final Applicant applicant) {
-
-        if (caseData.isDivorce()) {
-            log.info("Generating Judicial Separation Order Refused Cover Letter for case id {} ", caseId);
-        } else {
-            log.info("Generating Separation Order Refused Cover Letter for case id {} ", caseId);
-        }
-
-        caseDataDocumentService.renderDocumentAndUpdateCaseData(
-            caseData,
-            getCoverLetterDocumentType(caseData, applicant),
-            templateContent(caseData, caseId, applicant),
-            caseId,
-            getCoverLetterDocumentTemplateId(caseData, applicant),
-            applicant.getLanguagePreference(),
-            formatDocumentName(caseId, REJECTED_REFUSAL_ORDER_COVER_LETTER_DOCUMENT_NAME, now(clock))
-        );
-    }
-
-    public DocumentType getCoverLetterDocumentType(final CaseData caseData, final Applicant applicant) {
-        return conditionalOrderCommonContent.getCoverLetterDocumentType(caseData, applicant, false);
-    }
-
-    public String getCoverLetterDocumentTemplateId(final CaseData caseData, final Applicant applicant) {
-        return conditionalOrderCommonContent.getCoverLetterDocumentTemplateId(caseData, applicant, false);
-    }
-
-    private String getSolicitorName(final Applicant applicant) {
-        if (applicant.isRepresented()) {
-            return applicant.getSolicitor().getName();
-        }
-        return WELSH.equals(applicant.getLanguagePreference())
-            ? "nas cynrychiolwyd"
-            : "not represented";
-    }
-
-    private String getSolicitorReference(final Applicant applicant) {
-        if (applicant.isRepresented() && applicant.getSolicitor().getReference() != null) {
-            return applicant.getSolicitor().getReference();
-        }
-        return WELSH.equals(applicant.getLanguagePreference())
-            ? "heb ei ddarparu"
-            : "not provided";
-    }
 
     public Map<String, Object> templateContent(final CaseData caseData, final Long ccdCaseReference, final Applicant applicant) {
 
@@ -141,5 +86,23 @@ public class JudicialSeparationCoRefusalTemplateContent {
         }
 
         return templateContent;
+    }
+
+    private String getSolicitorName(final Applicant applicant) {
+        if (applicant.isRepresented()) {
+            return applicant.getSolicitor().getName();
+        }
+        return WELSH.equals(applicant.getLanguagePreference())
+            ? "nas cynrychiolwyd"
+            : "not represented";
+    }
+
+    private String getSolicitorReference(final Applicant applicant) {
+        if (applicant.isRepresented() && applicant.getSolicitor().getReference() != null) {
+            return applicant.getSolicitor().getReference();
+        }
+        return WELSH.equals(applicant.getLanguagePreference())
+            ? "heb ei ddarparu"
+            : "not provided";
     }
 }
