+++ src/test/java/uk/gov/hmcts/divorce/document/content/GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest.java
@@ -5,13 +5,13 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.Application;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.divorcecase.model.DivorceOrDissolution;
 import uk.gov.hmcts.divorce.divorcecase.model.Solicitor;
-import uk.gov.hmcts.divorce.document.CaseDataDocumentService;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 
 import java.time.Clock;
@@ -58,6 +58,8 @@ import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.SO
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.SOLICITOR_FIRM;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.SOLICITOR_NAME;
 import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.SOLICITOR_REFERENCE;
+import static uk.gov.hmcts.divorce.document.model.DocumentType.CONDITIONAL_ORDER_REFUSAL_COVER_LETTER;
+import static uk.gov.hmcts.divorce.document.model.DocumentType.JUDICIAL_SEPARATION_ORDER_CLARIFICATION_REFUSAL_SOLICITOR_COVER_LETTER;
 import static uk.gov.hmcts.divorce.document.model.DocumentType.JUDICIAL_SEPARATION_ORDER_REFUSAL_COVER_LETTER;
 import static uk.gov.hmcts.divorce.document.model.DocumentType.JUDICIAL_SEPARATION_ORDER_REFUSAL_SOLICITOR_COVER_LETTER;
 import static uk.gov.hmcts.divorce.notification.CommonContent.ADDRESS;
@@ -74,9 +76,6 @@ import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getBasicDocmosisTempl
 class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
 
     @Mock
-    private CaseDataDocumentService caseDataDocumentService;
-
-    @Mock
     private ConditionalOrderCommonContent conditionalOrderCommonContent;
 
     @Mock
@@ -89,7 +88,7 @@ class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
     private JudicialSeparationCoRefusalTemplateContent judicialSeparationCoRefusalTemplateContent;
 
     @Test
-    void shouldGenerateAndUpdateCaseData() {
+    void shouldFetchTemplateContentForJudicialSeparationCase() {
         setMockClock(clock);
 
         final List<ConditionalOrderCommonContent.RefusalReason> refusalReasons =
@@ -114,52 +113,30 @@ class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
             .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
             .applicationType(JOINT_APPLICATION)
             .supplementaryCaseType(JUDICIAL_SEPARATION)
-
-            .applicant1(
-                Applicant.builder()
+            .applicant1(Applicant.builder()
                     .firstName("Bob")
                     .lastName("Smith")
                     .address(APPLICANT_ADDRESS)
                     .languagePreferenceWelsh(NO)
-                    .build()
-            )
-            .applicant2(
-                Applicant.builder()
+                    .build())
+            .applicant2(Applicant.builder()
                     .gender(FEMALE)
-                    .build()
-            )
-            .conditionalOrder(
-                ConditionalOrder.builder()
+                    .build())
+            .conditionalOrder(ConditionalOrder.builder()
                     .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
-                    .build()
-            )
-            .build();
+                    .build()).build();
 
         templateContent.put(ADDRESS, caseData.getApplicant1().getPostalAddress());
 
-        when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
-            .thenReturn(refusalReasons);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentType(caseData, caseData.getApplicant1(), false))
-            .thenReturn(JUDICIAL_SEPARATION_ORDER_REFUSAL_COVER_LETTER);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentTemplateId(caseData, caseData.getApplicant1(), false))
-            .thenReturn(JUDICIAL_SEPARATION_ORDER_REFUSAL_COVER_LETTER_TEMPLATE_ID);
-        when(conditionalOrderCommonContent.getPartner(caseData))
-            .thenReturn("wife");
+        when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder())).thenReturn(refusalReasons);
+        when(conditionalOrderCommonContent.getPartner(caseData)).thenReturn("wife");
         when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
             caseData.getApplicant1().getLanguagePreference()));
 
-        judicialSeparationCoRefusalTemplateContent.generateAndUpdateCaseData(
-            caseData, TEST_CASE_ID, caseData.getApplicant1());
+        Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
+            TEST_CASE_ID, caseData.getApplicant1());
 
-        verify(caseDataDocumentService).renderDocumentAndUpdateCaseData(
-            eq(caseData),
-            eq(JUDICIAL_SEPARATION_ORDER_REFUSAL_COVER_LETTER),
-            eq(templateContent),
-            eq(TEST_CASE_ID),
-            eq(JUDICIAL_SEPARATION_ORDER_REFUSAL_COVER_LETTER_TEMPLATE_ID),
-            eq(ENGLISH),
-            anyString()
-        );
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
     }
 
     @Test
@@ -191,24 +168,19 @@ class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
             .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
             .applicationType(JOINT_APPLICATION)
             .supplementaryCaseType(JUDICIAL_SEPARATION)
-            .applicant1(
-                Applicant.builder()
+            .applicant1(Applicant.builder()
                     .solicitorRepresented(YES)
-                    .solicitor(
-                        Solicitor.builder()
+                    .solicitor(Solicitor.builder()
                             .name("App1 Solicitor Name")
                             .firmName("App1 Solicitor Firm")
                             .address("App1 Solicitor Address")
                             .reference("App1 SolicitorReference")
-                            .build()
-                    )
+                            .build())
                     .firstName("Bob")
                     .lastName("Smith")
                     .languagePreferenceWelsh(NO)
-                    .build()
-            )
-            .applicant2(
-                Applicant.builder()
+                    .build())
+            .applicant2(Applicant.builder()
                     .solicitorRepresented(YES)
                     .solicitor(
                         Solicitor.builder()
@@ -216,50 +188,27 @@ class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
                             .firmName("App2 Solicitor Firm")
                             .address("App2 Solicitor Address")
                             .reference("App2 SolicitorReference")
-                            .build()
-                    )
+                            .build())
                     .firstName("Roberta")
                     .lastName("Smith")
                     .build()
             )
-            .application(
-                Application.builder()
+            .application(Application.builder()
                     .newPaperCase(YES)
-                    .build()
-            )
-            .conditionalOrder(
-                ConditionalOrder.builder()
+                    .build())
+            .conditionalOrder(ConditionalOrder.builder()
                     .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
-                    .build()
-            )
-            .build();
+                    .build()).build();
 
         when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
             .thenReturn(refusalReasons);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentType(caseData, caseData.getApplicant1(), false))
-            .thenReturn(JUDICIAL_SEPARATION_ORDER_REFUSAL_SOLICITOR_COVER_LETTER);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentTemplateId(caseData, caseData.getApplicant1(), false))
-            .thenReturn(REJECTED_REFUSAL_ORDER_COVER_LETTER_TEMPLATE_ID);
         when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
             caseData.getApplicant1().getLanguagePreference()));
 
         Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
             TEST_CASE_ID, caseData.getApplicant1());
 
-        assertThat();
-
-        judicialSeparationCoRefusalTemplateContent.generateAndUpdateCaseData(
-            caseData, TEST_CASE_ID, caseData.getApplicant1());
-
-        verify(caseDataDocumentService).renderDocumentAndUpdateCaseData(
-            eq(caseData),
-            eq(JUDICIAL_SEPARATION_ORDER_REFUSAL_SOLICITOR_COVER_LETTER),
-            eq(templateContent),
-            eq(TEST_CASE_ID),
-            eq(REJECTED_REFUSAL_ORDER_COVER_LETTER_TEMPLATE_ID),
-            eq(ENGLISH),
-            anyString()
-        );
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
     }
 
     @Test
@@ -291,60 +240,212 @@ class GenerateJudicialSeparationCORefusedForAmendmentCoverLetterTest {
             .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
             .applicationType(SOLE_APPLICATION)
             .supplementaryCaseType(JUDICIAL_SEPARATION)
-            .applicant1(
-                Applicant.builder()
+            .applicant1(Applicant.builder()
                     .solicitorRepresented(YES)
-                    .solicitor(
-                        Solicitor.builder()
+                    .solicitor(Solicitor.builder()
                             .name("App1 Solicitor Name")
                             .firmName("App1 Solicitor Firm")
                             .address("App1 Solicitor Address")
-                            .build()
-                    )
+                            .build())
                     .firstName("Bob")
                     .lastName("Smith")
                     .languagePreferenceWelsh(NO)
-                    .build()
-            )
-            .applicant2(
-                Applicant.builder()
+                    .build())
+            .applicant2(Applicant.builder()
                     .solicitorRepresented(NO)
                     .firstName("Roberta")
                     .lastName("Smith")
-                    .build()
-            )
-            .application(
-                Application.builder()
+                    .build())
+            .application(Application.builder()
                     .newPaperCase(YES)
-                    .build()
-            )
+                    .build())
             .conditionalOrder(
                 ConditionalOrder.builder()
                     .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
-                    .build()
-            )
-            .build();
+                    .build()).build();
 
         when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
             .thenReturn(refusalReasons);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentType(caseData, caseData.getApplicant1(), false))
-            .thenReturn(JUDICIAL_SEPARATION_ORDER_REFUSAL_SOLICITOR_COVER_LETTER);
-        when(conditionalOrderCommonContent.getCoverLetterDocumentTemplateId(caseData, caseData.getApplicant1(), false))
-            .thenReturn(REJECTED_REFUSAL_ORDER_COVER_LETTER_TEMPLATE_ID);
         when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
             caseData.getApplicant1().getLanguagePreference()));
 
-        judicialSeparationCoRefusalTemplateContent.generateAndUpdateCaseData(
-            caseData, TEST_CASE_ID, caseData.getApplicant1());
+        Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
+            TEST_CASE_ID, caseData.getApplicant1());
 
-        verify(caseDataDocumentService).renderDocumentAndUpdateCaseData(
-            eq(caseData),
-            eq(JUDICIAL_SEPARATION_ORDER_REFUSAL_SOLICITOR_COVER_LETTER),
-            eq(templateContent),
-            eq(TEST_CASE_ID),
-            eq(REJECTED_REFUSAL_ORDER_COVER_LETTER_TEMPLATE_ID),
-            eq(ENGLISH),
-            anyString()
-        );
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
+    }
+
+    @Test
+    void shouldFetchTemplateContentForClarificationJudicialSeparationCase() {
+        setMockClock(clock);
+
+        final List<ConditionalOrderCommonContent.RefusalReason> refusalReasons =
+            List.of(new ConditionalOrderCommonContent.RefusalReason("Court does not have jurisdiction"));
+
+        final Map<String, Object> templateContent = new HashMap<>();
+        templateContent.put(CONTACT_EMAIL, CONTACT_DIVORCE_EMAIL);
+        templateContent.put(CASE_REFERENCE, formatId(TEST_CASE_ID));
+        templateContent.put(COURTS_AND_TRIBUNALS_SERVICE_HEADER, COURTS_AND_TRIBUNALS_SERVICE_HEADER_TEXT);
+        templateContent.put(DATE, LocalDate.now(clock).format(DATE_TIME_FORMATTER));
+        templateContent.put(IS_JOINT, true);
+        templateContent.put(FIRST_NAME, "Bob");
+        templateContent.put(LAST_NAME, "Smith");
+        templateContent.put(MARRIAGE_OR_CIVIL_PARTNERSHIP, MARRIAGE);
+        templateContent.put(DIVORCE_OR_CIVIL_PARTNERSHIP, CommonContent.DIVORCE);
+        templateContent.put(FEEDBACK, refusalReasons);
+        templateContent.put(PARTNER, "wife");
+        templateContent.put(DIVORCE_AND_DISSOLUTION_HEADER, DIVORCE_AND_DISSOLUTION_HEADER_TEXT);
+        templateContent.put(PHONE_AND_OPENING_TIMES, PHONE_AND_OPENING_TIMES_TEXT);
+
+        CaseData caseData = CaseData.builder()
+            .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
+            .applicationType(JOINT_APPLICATION)
+            .supplementaryCaseType(JUDICIAL_SEPARATION)
+            .applicant1(Applicant.builder()
+                    .firstName("Bob")
+                    .lastName("Smith")
+                    .address(APPLICANT_ADDRESS)
+                    .languagePreferenceWelsh(NO)
+                    .build())
+            .applicant2(Applicant.builder()
+                    .gender(FEMALE)
+                    .build())
+            .conditionalOrder(ConditionalOrder.builder()
+                    .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
+                    .build()).build();
+
+        templateContent.put(ADDRESS, caseData.getApplicant1().getPostalAddress());
+
+        when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
+            .thenReturn(refusalReasons);
+        when(conditionalOrderCommonContent.getPartner(caseData))
+            .thenReturn("wife");
+        when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
+            caseData.getApplicant1().getLanguagePreference()));
+
+        Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
+            TEST_CASE_ID, caseData.getApplicant1());
+
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
+    }
+
+    @Test
+    void shouldGenerateAndUpdateCaseDataForOfflineRepresentedApplicant() {
+        setMockClock(clock);
+
+        final List<ConditionalOrderCommonContent.RefusalReason> refusalReasons =
+            List.of(new ConditionalOrderCommonContent.RefusalReason("Court does not have jurisdiction"));
+
+        final Map<String, Object> templateContent = new HashMap<>();
+        templateContent.put(CONTACT_EMAIL, CONTACT_DIVORCE_EMAIL);
+        templateContent.put(CASE_REFERENCE, formatId(TEST_CASE_ID));
+        templateContent.put(COURTS_AND_TRIBUNALS_SERVICE_HEADER, COURTS_AND_TRIBUNALS_SERVICE_HEADER_TEXT);
+        templateContent.put(DATE, LocalDate.now(clock).format(DATE_TIME_FORMATTER));
+        templateContent.put(IS_JOINT, true);
+        templateContent.put(FEEDBACK, refusalReasons);
+        templateContent.put(DIVORCE_AND_DISSOLUTION_HEADER, DIVORCE_AND_DISSOLUTION_HEADER_TEXT);
+        templateContent.put(PHONE_AND_OPENING_TIMES, PHONE_AND_OPENING_TIMES_TEXT);
+        templateContent.put(SOLICITOR_NAME, "App1 Solicitor Name");
+        templateContent.put(SOLICITOR_FIRM, "App1 Solicitor Firm");
+        templateContent.put(SOLICITOR_ADDRESS, "App1 Solicitor Address");
+        templateContent.put(SOLICITOR_REFERENCE, "App1 SolicitorReference");
+        templateContent.put(APPLICANT_1_SOLICITOR_NAME, "App1 Solicitor Name");
+        templateContent.put(APPLICANT_2_SOLICITOR_NAME, "App2 Solicitor Name");
+        templateContent.put(APPLICANT_1_FULL_NAME, "Bob Smith");
+        templateContent.put(APPLICANT_2_FULL_NAME, "Roberta Smith");
+
+        CaseData caseData = CaseData.builder()
+            .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
+            .applicationType(JOINT_APPLICATION)
+            .supplementaryCaseType(JUDICIAL_SEPARATION)
+            .applicant1(Applicant.builder()
+                    .solicitorRepresented(YES)
+                    .solicitor(Solicitor.builder().name("App1 Solicitor Name").firmName("App1 Solicitor Firm")
+                        .address("App1 Solicitor Address").reference("App1 SolicitorReference").build())
+                    .firstName("Bob")
+                    .lastName("Smith")
+                    .languagePreferenceWelsh(NO)
+                    .build())
+            .applicant2(Applicant.builder()
+                    .solicitorRepresented(YES)
+                    .solicitor(
+                        Solicitor.builder().name("App2 Solicitor Name").firmName("App2 Solicitor Firm")
+                            .address("App2 Solicitor Address").reference("App2 SolicitorReference").build())
+                    .firstName("Roberta")
+                    .lastName("Smith")
+                    .build())
+            .application(Application.builder()
+                    .newPaperCase(YES)
+                    .build())
+            .conditionalOrder(
+                ConditionalOrder.builder()
+                    .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
+                    .build()).build();
+
+        when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
+            .thenReturn(refusalReasons);
+        when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
+            caseData.getApplicant1().getLanguagePreference()));
+
+        Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
+            TEST_CASE_ID, caseData.getApplicant1());
+
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
+    }
+
+    @Test
+    void shouldGenerateAndUpdateCaseDataForOfflineRepresentedApplicantWithNoReferenceAndUnrepresentedRespondent() {
+        setMockClock(clock);
+
+        final List<ConditionalOrderCommonContent.RefusalReason> refusalReasons =
+            List.of(new ConditionalOrderCommonContent.RefusalReason("Court does not have jurisdiction"));
+
+        final Map<String, Object> templateContent = new HashMap<>();
+        templateContent.put(CONTACT_EMAIL, CONTACT_DIVORCE_EMAIL);
+        templateContent.put(CASE_REFERENCE, formatId(TEST_CASE_ID));
+        templateContent.put(COURTS_AND_TRIBUNALS_SERVICE_HEADER, COURTS_AND_TRIBUNALS_SERVICE_HEADER_TEXT);
+        templateContent.put(DATE, LocalDate.now(clock).format(DATE_TIME_FORMATTER));
+        templateContent.put(IS_JOINT, false);
+        templateContent.put(FEEDBACK, refusalReasons);
+        templateContent.put(DIVORCE_AND_DISSOLUTION_HEADER, DIVORCE_AND_DISSOLUTION_HEADER_TEXT);
+        templateContent.put(PHONE_AND_OPENING_TIMES, PHONE_AND_OPENING_TIMES_TEXT);
+        templateContent.put(SOLICITOR_NAME, "App1 Solicitor Name");
+        templateContent.put(SOLICITOR_FIRM, "App1 Solicitor Firm");
+        templateContent.put(SOLICITOR_ADDRESS, "App1 Solicitor Address");
+        templateContent.put(SOLICITOR_REFERENCE, "not provided");
+        templateContent.put(APPLICANT_1_SOLICITOR_NAME, "App1 Solicitor Name");
+        templateContent.put(APPLICANT_2_SOLICITOR_NAME, "not represented");
+        templateContent.put(APPLICANT_1_FULL_NAME, "Bob Smith");
+        templateContent.put(APPLICANT_2_FULL_NAME, "Roberta Smith");
+
+        CaseData caseData = CaseData.builder()
+            .divorceOrDissolution(DivorceOrDissolution.DIVORCE)
+            .applicationType(SOLE_APPLICATION)
+            .supplementaryCaseType(JUDICIAL_SEPARATION)
+            .applicant1(Applicant.builder()
+                    .solicitorRepresented(YES)
+                    .solicitor(Solicitor.builder().name("App1 Solicitor Name").firmName("App1 Solicitor Firm")
+                            .address("App1 Solicitor Address").build())
+                    .firstName("Bob")
+                    .lastName("Smith")
+                    .languagePreferenceWelsh(NO)
+                    .build())
+            .applicant2(Applicant.builder()
+                    .solicitorRepresented(YesOrNo.NO)
+                    .firstName("Roberta").lastName("Smith").build())
+            .application(Application.builder().newPaperCase(YES).build())
+            .conditionalOrder(ConditionalOrder.builder()
+                    .refusalRejectionAdditionalInfo("Court does not have jurisdiction")
+                    .build()).build();
+
+        when(conditionalOrderCommonContent.generateLegalAdvisorComments(caseData.getConditionalOrder()))
+            .thenReturn(refusalReasons);
+        when(docmosisCommonContent.getBasicDocmosisTemplateContent(ENGLISH)).thenReturn(getBasicDocmosisTemplateContent(
+            caseData.getApplicant1().getLanguagePreference()));
+
+        Map<String, Object> actualTemplateContent = judicialSeparationCoRefusalTemplateContent.templateContent(caseData,
+            TEST_CASE_ID, caseData.getApplicant1());
+
+        assertThat(actualTemplateContent.entrySet()).containsAll(templateContent.entrySet());
     }
 }
